// src/app/features/manage-profile/edit-resume/personal/address/model/address-update.model.ts
export interface AddressDetails {
  guidId: string;
  presentAddressType: number;
  presentCountryValue: number;
  presentPostOfficeValue: number;
  presentAddressId: number;
  presentDistrictValue: number;
  presentThanaValue: number;
  presentVillageValue: string;
  presentIsInsert: number;
  PresentOutsideBd: number;

  permanentIsInsert: number;
  permanentAddressId: number;
  permanentDistrictValue: number;
  permanentThanaValue: number;
  permanentPostOfficeValue: number;
  permanentCountryValue: number;
  permanentVillageValue: string;
  PermanentOutsideBd : number;
  permanentAddressType: number;

}